pkg/listwatch: remove multilistwatcher
